A tree reader which reads a list of trees from a string or file.
Includes an interface to parse multiple trees from the same file.

Uses the sly module to implement a recursive descent parser.  It is
somewhat slow, but not so slow that it is a restriction in terms of
building a parser.  It is unclear whether the speed is because of the
sly module or because of our usage of the sly module.